# swagger.yaml
swagger: "2.0"
info:
  description: "PreKaraのバックエンドサーバーのAPIリファレンスです。ライセンス等はリポジトリに準拠します。/apiを叩くとAPIのバージョンが取得できます。"
  version: "0.0.1"
  title: "PreKara Server API.dev"
host: "local.prekara.mizucoffee.net"
basePath: "/api/v1"
tags:
- name: "server"
  description: "サーバー操作"
- name: "session"
  description: "セッション操作"
- name: "theme"
  description: "お題操作"
- name: "image"
  description: "画像操作"
- name: "control"
  description: "コントロール"
schemes:
- "http"
paths:
  /server:
    post:
      tags:
      - "server"
      summary: "新規サーバー"
      description: "サーバーを新規作成します"
      operationId: "newServer"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "サーバー情報"
        required: true
        schema:
          type: "object"
          required:
          - "server_name"
          - "password"
          properties:
            server_name:
              type: "string"
              example: "Server1"
            password:
              type: "string"
              example: "pass2"
      responses:
        200:
          description: "Success"
        409:
          description: "Conflict"
        405:
          description: "Invalid parameter"
    put:
      tags:
      - "server"
      summary: "サーバー情報変更"
      description: "bodyに'server_name'か'password'の片方かその両方が必要です"
      operationId: "updateServer"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "サーバー情報"
        required: true
        schema:
          type: "object"
          required:
          - "server_name"
          - "password"
          properties:
            server_name:
              type: "string"
              example: "Server2"
            password:
              type: "string"
              example: "pass2"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
        405:
          description: "Invalid parameter"
        409:
          description: "Conflict"
    delete:
      tags:
      - "server"
      summary: "サーバー削除"
      description: "同時にログアウト(セッションが破棄)されます"
      operationId: "deleteServer"
      produces:
      - "application/json"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
  /session:
    get:
      tags:
      - "session"
      summary: "セッション状態の取得"
      description: "Session情報を基に状態が返ります"
      operationId: "getSession"
      produces:
      - "application/json"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
    post:
      tags:
      - "session"
      summary: "セッションログイン"
      description: "ログインします。Cookieに記録されます。"
      operationId: "postSession"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          required:
          - "server_name"
          - "password"
          properties:
            server_name:
              type: "string"
              example: "Server1"
            password:
              type: "string"
              example: "pass1"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
        405:
          description: "Invalid parameter"
    delete:
      tags:
      - "session"
      summary: "セッション破棄"
      description: "セッションを破棄します。ログアウトと同義です。"
      operationId: "revokeSession"
      produces:
      - "application/json"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
  /theme:
    post:
      tags:
      - "theme"
      summary: "お題の投稿"
      operationId: "postTheme"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          required:
          - "theme"
          properties:
            theme:
              type: "string"
              example: "Theme1"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
        405:
          description: "Invalid parameter"
        409:
          description: "Conflict"
    delete:
      tags:
      - "theme"
      summary: "お題の削除"
      operationId: "deleteTheme"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          required:
          - "theme_id"
          properties:
            theme_id:
              type: "string"
              example: "123456"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
        405:
          description: "Invalid parameter"
  /theme/list:
    get:
      tags:
      - "theme"
      summary: "お題一覧の取得"
      operationId: "listTheme"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"

  /image:
    get:
      tags:
      - "image"
      summary: "画像の取得"
      operationId: "getImage"
      parameters:
      - in: "query"
        name: "image_name"
        type: "string"
        description: "取得する画像のファイル名"
        required: true
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
    post:
      tags:
      - "image"
      summary: "画像の転送"
      operationId: "postImage"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "image"
        in: "formData"
        description: "転送する画像"
        required: true
        type: "file"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
        405:
          description: "Invalid file"
    delete:
      tags:
      - "image"
      summary: "画像の削除"
      operationId: "deleteImage"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          required:
          - "image_name"
          properties:
            image_name:
              type: "string"
              example: "example.jpg"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
        405:
          description: "Invalid parameter"
  /image/list:
    get:
      tags:
      - "image"
      summary: "画像一覧の取得"
      operationId: "listImage"
      responses:
        200:
          description: "Success"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"